deploy/backend/ai-hub/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.28
pydantic>=2.0
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
groq>=0.9
httpx>=0.27
